agent: Handle both cgroup v1 and v2
Before this change, the `create_cgroups.sh` script was assuming that
cgroup hierarchy (mounted in `/sys/fs/cgroup`) is v1. The script was
also not exiting on error, so when running on a modern distro with
cgroups v2 (like Ubuntu 22.04), it was silently failing and then the
agent processes were running without cgroups at all.

This change fixes that by detecting the version of cgroup hierarchy and
applying changes differently for each version.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>