Configure file server URL
In case agents connect to console via a URL which is not used in browser, this can be changed to that URL for downloading threat intel feeds and uploading diagnostic logs.

Example: A domain name is configured in Okta and used to open UI. Agents connect via internal ip address or load balancer. Communication with file server can be changed to internal ip now.